05_python_powerpoint_generate/requirements.txt
python-multipart==0.0.20
fastapi 
uvicorn 
streamlit 
python-pptx 
pymupdf 
openai 
requests